Flask model/requirements.txt
Flask
Werkzeug
tensorflow
numpy 
Pillow